Moving svdpi.h into alternate location to make it clear that it's not ours.
